Jose branch

+++ C07051_B92291_Tarea_III_IF4101/Tarea3.Api/DTOs/ProductoDTO.cs
@@ -2,7 +2,7 @@
 {
     public class ProductoDTO
     {
-        public int idProducto { get; set; }
+        public long idProducto { get; set; }
         public string nombre { get; set; }
         public decimal precio { get; set; }
     }
