+++ C07051_B92291_Tarea_III_IF4101/Tarea3.DA/Entidades/ProductoDA.cs
@@ -13,7 +13,7 @@ namespace Tarea3.DA.Entidades
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.None)]
         [Required]
-        public int idProducto { get; set; }
+        public long idProducto { get; set; }
         [Required]
         public string nombre { get; set; }
         [Required]
